// test/integration/test/ports/engines/queries/data_types/bytes/_fixture/relations/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Child: {
      readonly childId: CodecTypes['pg/int4@1']['output'];
      readonly parentId: CodecTypes['pg/int4@1']['output'] | null;
      readonly bytes: CodecTypes['pg/bytea@1']['output'];
    };
    readonly Parent: { readonly id: CodecTypes['pg/int4@1']['output'] };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Child: {
      readonly childId: CodecTypes['pg/int4@1']['input'];
      readonly parentId: CodecTypes['pg/int4@1']['input'] | null;
      readonly bytes: CodecTypes['pg/bytea@1']['input'];
    };
    readonly Parent: { readonly id: CodecTypes['pg/int4@1']['input'] };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly child: {
      readonly bytes: CodecTypes['pg/bytea@1']['output'];
      readonly childId: CodecTypes['pg/int4@1']['output'];
      readonly parentId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly parent: { readonly id: CodecTypes['pg/int4@1']['output'] };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly child: {
      readonly bytes: CodecTypes['pg/bytea@1']['input'];
      readonly childId: CodecTypes['pg/int4@1']['input'];
      readonly parentId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly parent: { readonly id: CodecTypes['pg/int4@1']['input'] };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;